Organize world material
Though I may have gone a bit overboard with the subdividing and nesting into node groups, now anyone with a notion of cycles nodes and cycles node groups should be able to use it. Also made a simplified version for eevee to get higher framerate in the viewport.
Now if only there was a way to import cycles material into Unity :|